+++ PhuQuocVoucher.Api/Controllers/VoucherController.cs
@@ -1,7 +1,9 @@
-﻿using CrudApiTemplate.CustomException;
+﻿using CrudApiTemplate.CustomBinding;
+using CrudApiTemplate.CustomException;
 using CrudApiTemplate.Repository;
 using CrudApiTemplate.Request;
 using CrudApiTemplate.Utilities;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using PhuQuocVoucher.Business.Dtos.VoucherDto;
@@ -11,6 +13,7 @@ using PhuQuocVoucher.Data.Models;
 namespace PhuQuocVoucher.Api.Controllers;
 
 [ApiController]
+[Authorize(Roles = $"{nameof(Role.Admin)},{nameof(Role.Provider)}")]
 [Route("api/v1/[controller]s")]
 public class VoucherController : ControllerBase
 {
@@ -28,9 +31,9 @@ public class VoucherController : ControllerBase
     }
 
     [HttpGet("admin")]
-    public async Task<IActionResult> Get([FromQuery] FindVoucher request, [FromQuery] PagingRequest paging, string? orderBy)
+    public async Task<IActionResult> Get([FromQuery] FindVoucher request, [FromQuery] PagingRequest paging, string? orderBy,[FromClaim("ProviderId")] int? providerId)
     {
-        
+        request.ProviderId = providerId ?? request.ProviderId;
         return Ok((await _voucherService.GetAsync<VoucherView>(new GetRequest<Voucher>
         {
             FindRequest = request,
@@ -40,12 +43,11 @@ public class VoucherController : ControllerBase
     }
     
     
+    [AllowAnonymous]
     [HttpGet]
-    public async Task<IActionResult> GetDefault([FromQuery] FindVoucher request, [FromQuery] PagingRequest paging, string? orderBy)
+    public async Task<IActionResult> GetDefault([FromQuery] FindVoucher request, [FromQuery] PagingRequest paging, string? orderBy, [FromClaim("ProviderId")] int? providerId)
     {
-
-        request.Status = ModelStatus.Active;
-        
+        request.ProviderId = providerId ?? request.ProviderId;
         return Ok((await _voucherService.GetAsync<VoucherView>(new GetRequest<Voucher>
         {
             FindRequest = request,
@@ -55,32 +57,43 @@ public class VoucherController : ControllerBase
     }
 
     [HttpPost]
-    public async Task<IActionResult> Create([FromBody] CreateVoucher request)
+    public async Task<IActionResult> Create([FromBody] CreateVoucher request, [FromClaim("ProviderId")] int? providerId)
     {
+        request.ProviderId = providerId ?? request.ProviderId;
         return Ok(await _voucherService.CreateAsync(request));
     }
 
     [HttpGet("{id:int}")]
-    public async Task<IActionResult> Get(int id)
+    public async Task<IActionResult> Get(int id, [FromClaim("ProviderId")] int? providerId)
     {
-        return Ok(await _repo.Find<VoucherView>(cus => cus.Id == id).FirstOrDefaultAsync() ??
+        return Ok(await _repo.Find<VoucherView>(cus => cus.Id == id && (providerId == null || cus.ProviderId == providerId) ).FirstOrDefaultAsync() ??
                   throw new ModelNotFoundException($"Not Found {nameof(Voucher)} with id {id}"));
     }
 
     [HttpPut("{id:int}")]
-    public async Task<IActionResult> Update([FromBody] UpdateVoucher request, int id)
+    public async Task<IActionResult> Update([FromBody] UpdateVoucher request, int id, [FromClaim("ProviderId")] int? providerId)
     {
-        return Ok(await _voucherService.UpdateAsync(id, request));
+        var voucherIds = await _repo.Find(v => v.ProviderId == providerId).Select(v => v.Id).ToListAsync();
+        if (voucherIds.Contains(id))
+            return Ok(await _voucherService.UpdateAsync(id, request));
+        
+        return Forbid("Invalid Access");
     }
     [HttpPut("{id:int}/tag")]
-    public async Task<IActionResult> UpdateTags([FromBody] IList<int> tags, int id)
+    public async Task<IActionResult> UpdateTags([FromBody] IList<int> tags, int id, [FromClaim("ProviderId")] int? providerId)
     {
-        return Ok(await _voucherService.UpdateTag(tags, id));
+        var voucherIds = await _repo.Find(v => v.ProviderId == providerId).Select(v => v.Id).ToListAsync();
+        if (voucherIds.Contains(id))
+            return Ok(await _voucherService.UpdateTag(tags, id));
+        return Forbid("Invalid Access");
     }
     
     [HttpDelete("{id:int}")]
-    public async Task<IActionResult> Delete(int id)
-    {
-        return Ok(await _voucherService.DeleteAsync(id));
+    public async Task<IActionResult> Delete(int id, [FromClaim("ProviderId")] int? providerId)
+    {var voucherIds = await _repo.Find(v => v.ProviderId == providerId).Select(v => v.Id).ToListAsync();
+        if (voucherIds.Contains(id))
+            return Ok(await _voucherService.DeleteAsync(id));;
+        return Forbid("Invalid Access");
     }
+    
 }
